gui: improve `signers` and `signer` screens

+++ crates/coinstr/src/app/screen/signer.rs
@@ -1,22 +1,28 @@
 // Copyright (c) 2022-2023 Coinstr
 // Distributed under the MIT software license
 
-use coinstr_sdk::core::signer::Signer;
+use std::collections::BTreeMap;
+
+use coinstr_sdk::core::bitcoin::XOnlyPublicKey;
+use coinstr_sdk::core::signer::{Signer, SignerType};
 use coinstr_sdk::nostr::EventId;
 use coinstr_sdk::util;
 use iced::widget::{Column, Row, Space};
-use iced::{Command, Element, Length};
+use iced::{Alignment, Command, Element, Length};
 
 use crate::app::component::Dashboard;
 use crate::app::{Context, Message, Stage, State};
-use crate::component::{button, Text};
+use crate::component::{button, rule, Text};
 use crate::constants::APP_NAME;
 use crate::theme::color::RED;
 use crate::theme::icon::TRASH;
 
 #[derive(Debug, Clone)]
 pub enum SignerMessage {
+    LoadMySharedSigners(BTreeMap<EventId, XOnlyPublicKey>),
     Delete,
+    RevokeSharedSigner(EventId),
+    Reload,
     ErrorChanged(Option<String>),
 }
 
@@ -26,6 +32,7 @@ pub struct SignerState {
     loaded: bool,
     signer_id: EventId,
     signer: Signer,
+    my_shared_signers: BTreeMap<EventId, XOnlyPublicKey>,
     error: Option<String>,
 }
 
@@ -36,6 +43,7 @@ impl SignerState {
             loaded: false,
             signer_id,
             signer,
+            my_shared_signers: BTreeMap::new(),
             error: None,
         }
     }
@@ -49,12 +57,14 @@ impl State for SignerState {
         )
     }
 
-    fn load(&mut self, _ctx: &Context) -> Command<Message> {
-        /* let client = ctx.client.clone();
-        let signer_id = self.signer_id; */
-        self.loading = false;
-        self.loaded = true;
-        Command::none()
+    fn load(&mut self, ctx: &Context) -> Command<Message> {
+        self.loading = true;
+        let client = ctx.client.clone();
+        let signer_id = self.signer_id;
+        Command::perform(
+            async move { client.get_my_shared_signers(signer_id).unwrap() },
+            |signers| SignerMessage::LoadMySharedSigners(signers).into(),
+        )
     }
 
     fn update(&mut self, ctx: &mut Context, message: Message) -> Command<Message> {
@@ -64,22 +74,41 @@ impl State for SignerState {
 
         if let Message::Signer(msg) = message {
             match msg {
+                SignerMessage::LoadMySharedSigners(signers) => {
+                    self.my_shared_signers = signers;
+                    self.loading = false;
+                    self.loaded = true;
+                }
                 SignerMessage::ErrorChanged(error) => {
                     self.loading = false;
                     self.error = error;
                 }
                 SignerMessage::Delete => {
+                    if self.signer.signer_type() != SignerType::Seed {
+                        self.loading = true;
+                        let client = ctx.client.clone();
+                        let signer_id = self.signer_id;
+                        return Command::perform(
+                            async move { client.delete_signer_by_id(signer_id, None).await },
+                            |res| match res {
+                                Ok(_) => Message::View(Stage::Signers),
+                                Err(e) => SignerMessage::ErrorChanged(Some(e.to_string())).into(),
+                            },
+                        );
+                    }
+                }
+                SignerMessage::RevokeSharedSigner(shared_signer_id) => {
                     self.loading = true;
                     let client = ctx.client.clone();
-                    let signer_id = self.signer_id;
                     return Command::perform(
-                        async move { client.delete_signer_by_id(signer_id, None).await },
+                        async move { client.revoke_shared_signer(shared_signer_id).await },
                         |res| match res {
-                            Ok(_) => Message::View(Stage::Signers),
+                            Ok(_) => SignerMessage::Reload.into(),
                             Err(e) => SignerMessage::ErrorChanged(Some(e.to_string())).into(),
                         },
                     );
                 }
+                SignerMessage::Reload => return self.load(ctx),
             }
         }
 
@@ -111,7 +140,7 @@ impl State for SignerState {
 
             let mut delete_btn = button::danger_with_icon(TRASH, "Delete");
 
-            if !self.loading {
+            if !self.loading && self.signer.signer_type() != SignerType::Seed {
                 delete_btn = delete_btn.on_press(SignerMessage::Delete.into());
             }
 
@@ -124,7 +153,52 @@ impl State for SignerState {
                 content = content.push(Text::new(error).color(RED).view());
             };
 
-            // TODO: show shared signers (nostr pubkey, timestamp and revoke btn)
+            if !self.my_shared_signers.is_empty() {
+                content = content
+                    .push(Text::new("My Shared Signers").bold().bigger().view())
+                    .push(Space::with_height(10.0))
+                    .push(
+                        Row::new()
+                            .push(
+                                Text::new("ID")
+                                    .bold()
+                                    .bigger()
+                                    .width(Length::Fixed(115.0))
+                                    .view(),
+                            )
+                            .push(Text::new("User").bold().bigger().width(Length::Fill).view())
+                            .push(Space::with_width(Length::Fixed(40.0)))
+                            .spacing(10)
+                            .align_items(Alignment::Center)
+                            .width(Length::Fill),
+                    )
+                    .push(rule::horizontal_bold());
+
+                for (shared_signer_id, public_key) in self.my_shared_signers.iter() {
+                    let row = Row::new()
+                        .push(
+                            Text::new(util::cut_event_id(*shared_signer_id))
+                                .width(Length::Fixed(115.0))
+                                .view(),
+                        )
+                        .push(
+                            Text::new(util::cut_public_key(*public_key))
+                                .width(Length::Fill)
+                                .view(),
+                        )
+                        .push(
+                            button::danger_border_only_icon(TRASH)
+                                .on_press(
+                                    SignerMessage::RevokeSharedSigner(*shared_signer_id).into(),
+                                )
+                                .width(Length::Fixed(40.0)),
+                        )
+                        .spacing(10)
+                        .align_items(Alignment::Center)
+                        .width(Length::Fill);
+                    content = content.push(row).push(rule::horizontal());
+                }
+            }
         } else {
             content = content.push(Text::new("Loading...").view())
         };
